Check if loading prebuilt is allowed
