+++ .github/workflows/ci.yml
@@ -78,6 +78,31 @@ jobs:
       - name: run tests
         run: cargo test -p scarb-metadata
 
+  test-prebuilt-plugins:
+    name: test prebuilt plugins ${{ matrix.platform.name }}
+    runs-on: ${{ matrix.platform.os }}
+    # This is isolated, so it can be run on more platforms.
+    strategy:
+      fail-fast: false
+      matrix:
+        platform:
+          # Linux and windows are tested as part of normal CI run.
+          #- name: linux x86-64
+          #  os: ubuntu-latest
+          #- name: windows x86-64
+          #  os: windows-latest
+          - name: macos arm64
+            os: macos-latest
+          - name: macos x86-64
+            os: macos-13
+    steps:
+      - uses: actions/checkout@v4
+      - uses: dtolnay/rust-toolchain@stable
+      - uses: Swatinem/rust-cache@v2
+      - name: Run prebuilt plugin tests
+        run: |
+          cargo test -p scarb --test proc_macro_prebuilt
+
   check-rust:
     runs-on: ubuntu-latest
     steps:
@@ -108,29 +133,3 @@ jobs:
           cache-dependency-path: website/package-lock.json
       - run: npm ci
       - run: npm run fmt:check
-
-  test-prebuilt-plugins:
-    name: test prebuilt plugins ${{ matrix.platform.name }}
-    runs-on: ${{ matrix.platform.os }}
-    strategy:
-      fail-fast: false
-      matrix:
-        platform:
-          - name: linux x86-64
-            os: ubuntu-latest
-          - name: windows x86-64
-            os: windows-latest
-          - name: macos arm64
-            os: macos-latest
-          - name: macos x86-64
-            os: macos-13
-
-    steps:
-      - uses: actions/checkout@v4
-      - uses: dtolnay/rust-toolchain@stable
-      - uses: Swatinem/rust-cache@v2
-      - name: Run prebuilt plugin tests
-        run: |
-          cargo test -p scarb --test build_cairo_plugin compile_with_prebuilt_plugins -- --exact
-          cargo test -p scarb --test build_cairo_plugin compile_with_invalid_prebuilt_plugins -- --exact
-          cargo test -p scarb --test proc_macro_server load_prebuilt_proc_macros -- --exact
